api/requirements.txt
Django==3.2.4
django-cors-headers==3.7.0
django-crontab==0.7.1
django-filter==2.4.0
django-redis==5.0.0
djangorestframework==3.12.4
eventregistry==8.9
marshmallow==3.12.1
Markdown==3.3.4
martor==1.6.4
psycopg2-binary==2.9.1
pytest==6.2.4
pytz==2021.1
redis==3.5.3
gunicorn==20.1.0
favicon==0.7.0
requests==2.25.1
sentry-sdk==1.1.0
